Factoring away a simple inheritance form
The original form was done as part of a naive solution. The other form
was there based on how forms are looked up.

I'm removing the naive solution in preference for the more normalized
solution.
